Create LinkedListFunctions.c
Program containing different functions relating to manipulation of linked lists. Includes a function that prints out all the data in the linked list with only int data using recursion and not using recursion, and implementation a linked list (with print and append) that can store objects of type type int, float, and double.